NoMappingFoundException moved to IO package
